requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
google-genai
typer
